backend/pyproject.toml
[project]
name = "pdf-classaction-rag"
version = "0.1.0"
description = "RAG pipeline for PDF class action documents"
requires-python = ">=3.14"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "pydantic>=2.0.0",
    "pymupdf>=1.24.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "anthropic>=0.39.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pytesseract>=0.3.13",
    "reductoai>=0.16.0",
    "tabulate>=0.9.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]
reducto = [
    "reductoai>=0.16.0",
]
rerank-cohere = [
    "cohere>=5.0.0",
]
rerank-local = [
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_llm_server"]
